requirements.txt
drf-yasg
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.25
django-ckeditor==6.7.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
djangorestframework==3.16.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
packaging==25.0
pillow==11.3.0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
sqlparse==0.5.3
swagger-ui-py==25.7.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
